foundation/aistudio-infra/scripts/requirements.txt
azure-identity 
azure-ai-generative[faiss] 
azure-search-documents 
azure-ai-resources
azure-ai-ml
promptflow 
promptflow-tools
keyrings.alt